+++ templates/template-custom.php
@@ -1,3 +1,9 @@
+<?php
+/**
+ * Template Name: Custom Template
+ */
+?>
+
 <?php while (have_posts()) : the_post(); ?>
   <?php get_template_part('templates/partials/page', 'header'); ?>
   <?php get_template_part('templates/contents/content', 'page'); ?>